+++ it/controllers/SessionExpiredControllerISpec.scala
@@ -1,0 +1,111 @@
+/*
+ * Copyright 2021 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package controllers
+
+import org.jsoup.Jsoup
+import org.jsoup.nodes.Document
+import play.api.http.Status.OK
+import play.api.libs.ws.WSResponse
+import utils.{IntegrationTest, ViewHelpers}
+
+class SessionExpiredControllerISpec extends IntegrationTest with ViewHelpers {
+
+  object ExpectedResults {
+    object ContentEN {
+      val h1Expected = "For your security, we signed you out"
+      val p1Expected = "We did not save your answers."
+      val buttonExpectedText = "Sign in"
+      val buttonExpectedUrl: String = "http://localhost:11111/income-through-software/return/2022/start"
+    }
+
+    object ContentCY {
+      val h1Expected = "For your security, we signed you out"
+      val p1Expected = "We did not save your answers."
+      val buttonExpectedText = "Sign in"
+      val buttonExpectedUrl: String = "http://localhost:11111/income-through-software/return/2022/start"
+    }
+  }
+
+  object Selectors {
+    val h1Selector = "#main-content > div > div > header > h1"
+    val p1Selector = "#main-content > div > div > div.govuk-body > p"
+    val buttonSelector = "#continue"
+    val formSelector = "#main-content > div > div > form"
+  }
+
+  val url = s"${appUrl(port)}/timeout"
+
+  "When set to english" when {
+
+    import ExpectedResults.ContentEN._
+    import Selectors._
+
+    "the page is requested" should {
+
+      "render the page" which {
+        lazy val result: WSResponse = {
+          authoriseIndividual()
+          urlGet(url)
+        }
+
+        implicit def document: () => Document = () => Jsoup.parse(result.body)
+
+        "has an OK status" in {
+          result.status shouldBe OK
+        }
+
+        titleCheck(h1Expected)
+        welshToggleCheck("English")
+        h1Check(h1Expected, "xl")
+
+        textOnPageCheck(p1Expected,p1Selector)
+        buttonCheck(buttonExpectedText, buttonSelector)
+        formGetLinkCheck(buttonExpectedUrl, formSelector)
+      }
+    }
+  }
+
+  "When set to welsh" when {
+
+    import ExpectedResults.ContentCY._
+    import Selectors._
+
+    "the page is requested" should {
+
+      "render the page" which {
+        lazy val result: WSResponse = {
+          authoriseIndividual()
+          urlGet(url, true)
+        }
+
+        implicit def document: () => Document = () => Jsoup.parse(result.body)
+
+        "has an OK status" in {
+          result.status shouldBe OK
+        }
+
+        titleCheck(h1Expected)
+        welshToggleCheck("Welsh")
+        h1Check(h1Expected, "xl")
+
+        textOnPageCheck(p1Expected,p1Selector)
+        buttonCheck(buttonExpectedText, buttonSelector)
+        formGetLinkCheck(buttonExpectedUrl, formSelector)
+      }
+    }
+  }
+}
